feat: implement system hardening, agy CLI integration, and unified graph projection (T44-T50)
- Added ultra-fast async daemon handshake (<10ms) and CLI fast-fail.

- Implemented structured NDJSON error contracts for better automation.

- Integrated agy CLI transcript capture with real-time hooks and multi-path discovery.

- Refactored graph layer to support unified, buffered projection to both SQLite and CozoDB.

- Hardened capture privacy and enforced strict Rust safety across all new paths.

- Verified with 170+ tests and cross-model Codex review.

+++ crates/ai-brains-cli/src/commands/safety.rs
@@ -4,8 +4,8 @@ use serde::Deserialize;
 struct ChangeGuardHotspot {
     path: String,
     score: f64,
-    complexity: u32,
-    frequency: u32,
+    complexity: f64,
+    frequency: f64,
 }
 
 pub fn run(
@@ -148,8 +148,8 @@ fn fetch_hotspots_text(limit: usize) -> Result<Vec<ChangeGuardHotspot>, String> 
         if parts.len() >= 6 {
             if let (Ok(score), Ok(frequency), Ok(complexity)) = (
                 parts[2].parse::<f64>(),
-                parts[3].parse::<u32>(),
-                parts[4].parse::<u32>(),
+                parts[3].parse::<f64>(),
+                parts[4].parse::<f64>(),
             ) {
                 let path = parts[5].to_string();
                 if !path.is_empty() && path != "File Path" {
